Remove BaseChallengeContext and BaseSignOutContext
